refactor(file-links): fold trim into #15240 shared text-boundary module
Move non-ASCII prose boundary from terminal-http-url-extraction into
src/shared/non-ascii-terminal-text-boundary.ts and wire file-link trim
through the same module (#15240 mirror).
