+++ src/renderer/src/components/terminal-pane/terminal-http-url-extraction.ts
@@ -1,4 +1,5 @@
 import { TERMINAL_HTTP_URL_MAX_LENGTH } from './terminal-http-link-limits'
+import { isNonAsciiProseBoundary } from '../../../../shared/non-ascii-terminal-text-boundary'
 
 type ParsedTerminalHttpLink = {
   url: string
@@ -98,7 +99,7 @@ function trimHttpUrlTrailingPunctuation(
 
 function isHttpUrlBodyTerminator(code: number): boolean {
   return (
-    isNonAsciiUrlTerminator(code) ||
+    isNonAsciiProseBoundary(code) ||
     isAsciiWhitespace(code) ||
     code === 0x22 ||
     code === 0x27 ||
@@ -141,14 +142,6 @@ function isHttpUrlTrailingPunctuation(code: number): boolean {
     code === 0x3e ||
     code === 0x60
   )
-}
-
-// Why not all non-ascii: a url path may legitimately carry unencoded CJK (see the wrapped-url
-// tests), so only punctuation and spaces end it - full-width brackets, an ideographic space, `\u30fb`.
-const NON_ASCII_URL_TERMINATOR = /[\p{P}\p{S}\p{Z}]/u
-
-function isNonAsciiUrlTerminator(code: number): boolean {
-  return code > 0x7e && NON_ASCII_URL_TERMINATOR.test(String.fromCharCode(code))
 }
 
 function isAsciiWhitespace(code: number): boolean {
